+++ frb_codegen/src/library/codegen/parser/mir/parser/attribute.rs
@@ -532,7 +532,7 @@ impl Parse for FrbAttributeName {
 #[cfg(test)]
 mod tests {
     use crate::codegen::ir::mir::default::MirDefaultValue;
-    use crate::codegen::parser::mir::attribute::{
+    use crate::codegen::parser::mir::parser::attribute::{
         FrbAttribute, FrbAttributeDartCode, FrbAttributeDefaultValue, FrbAttributeMirror,
         FrbAttributeName, FrbAttributes, NamedOption,
     };
