+++ frb_codegen/src/library/codegen/parser/mir/parser/ty/concrete.rs
@@ -10,9 +10,9 @@ use crate::codegen::ir::mir::ty::dynamic::MirTypeDynamic;
 use crate::codegen::ir::mir::ty::general_list::mir_list;
 use crate::codegen::ir::mir::ty::MirType;
 use crate::codegen::ir::mir::ty::MirType::{Boxed, DartOpaque, Delegate, Dynamic};
-use crate::codegen::parser::mir::type_parser::path_data::extract_path_data;
-use crate::codegen::parser::mir::type_parser::unencodable::{splay_segments, SplayedSegment};
-use crate::codegen::parser::mir::type_parser::TypeParserWithContext;
+use crate::codegen::parser::mir::parser::ty::path_data::extract_path_data;
+use crate::codegen::parser::mir::parser::ty::unencodable::{splay_segments, SplayedSegment};
+use crate::codegen::parser::mir::parser::ty::TypeParserWithContext;
 use crate::if_then_some;
 use anyhow::{bail, Context};
 use itertools::Itertools;
