+++ frb_codegen/src/library/codegen/parser/mir/parser/ty/dart_fn.rs
@@ -2,8 +2,8 @@ use crate::codegen::ir::mir::ty::dart_fn::MirDartFnOutput;
 use crate::codegen::ir::mir::ty::dart_fn::MirTypeDartFn;
 use crate::codegen::ir::mir::ty::delegate::MirTypeDelegate;
 use crate::codegen::ir::mir::ty::MirType;
-use crate::codegen::parser::mir::type_parser::result::{parse_type_maybe_result, ResultTypeInfo};
-use crate::codegen::parser::mir::type_parser::TypeParserWithContext;
+use crate::codegen::parser::mir::parser::ty::result::{parse_type_maybe_result, ResultTypeInfo};
+use crate::codegen::parser::mir::parser::ty::TypeParserWithContext;
 use crate::if_then_some;
 use anyhow::{bail, Context};
 use syn::{
