+++ frb_codegen/src/library/codegen/parser/mir/parser/ty/enumeration.rs
@@ -11,14 +11,14 @@ use crate::codegen::ir::mir::ty::rust_auto_opaque_implicit::MirTypeRustAutoOpaqu
 use crate::codegen::ir::mir::ty::structure::MirStruct;
 use crate::codegen::ir::mir::ty::MirType;
 use crate::codegen::ir::mir::ty::MirType::{Delegate, EnumRef};
-use crate::codegen::parser::mir::attribute::FrbAttributes;
-use crate::codegen::parser::mir::type_parser::enum_or_struct::{
+use crate::codegen::parser::mir::parser::attribute::FrbAttributes;
+use crate::codegen::parser::mir::parser::ty::enum_or_struct::{
     parse_struct_or_enum_should_ignore, EnumOrStructParser, EnumOrStructParserInfo,
 };
-use crate::codegen::parser::mir::type_parser::misc::parse_comments;
-use crate::codegen::parser::mir::type_parser::structure::structure_compute_default_opaque;
-use crate::codegen::parser::mir::type_parser::unencodable::SplayedSegment;
-use crate::codegen::parser::mir::type_parser::TypeParserWithContext;
+use crate::codegen::parser::mir::parser::ty::misc::parse_comments;
+use crate::codegen::parser::mir::parser::ty::structure::structure_compute_default_opaque;
+use crate::codegen::parser::mir::parser::ty::unencodable::SplayedSegment;
+use crate::codegen::parser::mir::parser::ty::TypeParserWithContext;
 use crate::if_then_some;
 use crate::utils::namespace::{Namespace, NamespacedName};
 use std::collections::HashMap;
