+++ frb_codegen/src/library/codegen/parser/mir/parser/ty/mod.rs
@@ -10,11 +10,11 @@ use crate::codegen::ir::mir::ty::rust_opaque::RustOpaqueCodecMode;
 use crate::codegen::ir::mir::ty::structure::{MirStruct, MirStructIdent};
 use crate::codegen::ir::mir::ty::MirContext;
 use crate::codegen::ir::mir::ty::MirType;
-use crate::codegen::parser::mir::attribute::FrbAttributes;
-use crate::codegen::parser::mir::type_parser::array::ArrayParserInfo;
-use crate::codegen::parser::mir::type_parser::enum_or_struct::EnumOrStructParserInfo;
-use crate::codegen::parser::mir::type_parser::rust_auto_opaque_implicit::RustAutoOpaqueParserInfo;
-use crate::codegen::parser::mir::type_parser::rust_opaque::RustOpaqueParserInfo;
+use crate::codegen::parser::mir::parser::attribute::FrbAttributes;
+use crate::codegen::parser::mir::parser::ty::array::ArrayParserInfo;
+use crate::codegen::parser::mir::parser::ty::enum_or_struct::EnumOrStructParserInfo;
+use crate::codegen::parser::mir::parser::ty::rust_auto_opaque_implicit::RustAutoOpaqueParserInfo;
+use crate::codegen::parser::mir::parser::ty::rust_opaque::RustOpaqueParserInfo;
 use crate::utils::namespace::Namespace;
 use std::collections::HashMap;
 use syn::Type;
