+++ frb_codegen/src/library/codegen/parser/mir/parser/ty/result.rs
@@ -1,8 +1,8 @@
 use crate::codegen::ir::mir::ty::delegate::MirTypeDelegate;
 use crate::codegen::ir::mir::ty::MirType;
 use crate::codegen::ir::mir::ty::MirType::{EnumRef, StructRef};
-use crate::codegen::parser::mir::type_parser::unencodable::splay_segments;
-use crate::codegen::parser::mir::type_parser::{TypeParser, TypeParserParsingContext};
+use crate::codegen::parser::mir::parser::ty::unencodable::splay_segments;
+use crate::codegen::parser::mir::parser::ty::{TypeParser, TypeParserParsingContext};
 use anyhow::Context;
 
 #[allow(clippy::single_match)] // deliberate do so to ensure style consistency
