+++ frb_codegen/src/library/codegen/parser/mir/parser/ty/rust_auto_opaque_implicit.rs
@@ -6,11 +6,11 @@ use crate::codegen::ir::mir::ty::rust_opaque::{
     MirRustOpaqueInner, MirTypeRustOpaque, RustOpaqueCodecMode,
 };
 use crate::codegen::ir::mir::ty::{MirType, MirTypeTrait};
-use crate::codegen::parser::mir::type_parser::path_data::extract_path_data;
-use crate::codegen::parser::mir::type_parser::rust_opaque::{
+use crate::codegen::parser::mir::parser::ty::path_data::extract_path_data;
+use crate::codegen::parser::mir::parser::ty::rust_opaque::{
     GeneralizedRustOpaqueParserInfo, RustOpaqueParserTypeInfo,
 };
-use crate::codegen::parser::mir::type_parser::TypeParserWithContext;
+use crate::codegen::parser::mir::parser::ty::TypeParserWithContext;
 use crate::utils::namespace::Namespace;
 use anyhow::Result;
 use quote::ToTokens;
