+++ frb_codegen/src/library/codegen/parser/mir/parser/ty/slice.rs
@@ -1,6 +1,6 @@
 use crate::codegen::ir::mir::ty::general_list::mir_list;
 use crate::codegen::ir::mir::ty::MirType;
-use crate::codegen::parser::mir::type_parser::TypeParserWithContext;
+use crate::codegen::parser::mir::parser::ty::TypeParserWithContext;
 
 impl<'a, 'b, 'c> TypeParserWithContext<'a, 'b, 'c> {
     pub(crate) fn parse_type_slice(
